Project Control refactoring
